Unit tests for raw keyrings
